Implement displaying latlong position while stationary.
The latlong position is shown when the speed falls below the threshold
value of 5 MPH / 8 KPH.

The startup version string is now displayed for 5 seconds (up from 2 seconds)

Removed a lot of old "debug" cruft which doodled on the display at the same time
as trying to display the regular content.  The display isn't big enough for that.

Discovered that the display.display() was displaying only every other frame.
This isn't seen in the OLED demonstration programs.
For the moment, the "FIXME" hack is to call display.display() twice.
Found this trying to display the version message only once.
This also makes the waiting for GPS animation a lot smoother!

Other things
- Added a QR code, but it's not that usable at present.